refactor(agent): extract dashboard event callbacks into agent/events.py
Separate LangChain callback event translation from the run_turn primitive so the turn runner and graph-backed CLI path can share the same dashboard event shapes without coupling either driver to the other.
